[FEAT] 개인 소비 지출 입력 API (#85)

+++ backend/src/main/java/com/moong/domain/entity/GroupExpense.java
@@ -41,7 +41,7 @@ public class GroupExpense {
     @NotNull
     private String usage;
 
-    private int cost;
+    private long cost;
 
     @NotNull
     @ManyToOne(fetch = FetchType.LAZY)
